requirements.txt
lark-parser
pyyaml
pytest
pytest-cov